IS-3269: Flyttet metoder fra GraphApiService til GraphApiClient
