+++ src/test/kotlin/no/nav/syfo/testhelper/ExternalMockEnvironment.kt
@@ -2,7 +2,6 @@ package no.nav.syfo.testhelper
 
 import no.nav.syfo.application.ApplicationState
 import no.nav.syfo.application.cache.ValkeyStore
-import no.nav.syfo.testhelper.mock.GraphApiServiceMock
 import no.nav.syfo.testhelper.mock.getMockHttpClient
 import no.nav.syfo.testhelper.mock.wellKnownInternalAzureAD
 import redis.clients.jedis.DefaultJedisClientConfig
@@ -14,7 +13,6 @@ class ExternalMockEnvironment() {
     val applicationState: ApplicationState = testAppState()
     val environment = testEnvironment()
     val mockHttpClient = getMockHttpClient(env = environment)
-    val mockGraphApiService = GraphApiServiceMock()
 
     val redisConfig = environment.valkeyConfig
     val valkeyCache = ValkeyStore(
